Move core data model into public API where it should be.
